requirements.txt
feedparser>=6.0.11
groq>=0.11.0
edge-tts>=6.1.9
python-dotenv>=1.0.0
pillow>=10.0.0
